Removed serializer and old model files (#288)
* Removed serializer.ts

* Updated references to the EditorModel
